+++ src/extension.ts
@@ -17,16 +17,47 @@ export function activate(context: vscode.ExtensionContext) {
     let counter = new LineCount(context);
     context.subscriptions.push(counter);
 
-    let disposable1 = vscode.commands.registerCommand('extension.linecount.currentfile', () => {
-        counter.countCurrentFile();
-    });
+    let status = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left,-100);
+    status.text = 'LineCount';
+    status.command = 'extension.linecount.showcommands';
+    context.subscriptions.push(status);
+
+    let disposConfig = vscode.workspace.onDidChangeConfiguration(configChanged);
+    context.subscriptions.push(disposConfig);
+
+    let disposable1 = vscode.commands.registerCommand('extension.linecount.currentfile', countCurrentFile);
     context.subscriptions.push(disposable1);
 
-    let disposable2 = vscode.commands.registerCommand('extension.linecount.workspace', () => {
-        counter.countWorkspace();      
-    });
+    let disposable2 = vscode.commands.registerCommand('extension.linecount.workspace', countWorkspace);
     context.subscriptions.push(disposable2);
     
+    let disposable3 = vscode.commands.registerCommand('extension.linecount.showcommands', () => {
+        let items = [];
+        items.push({label: "LineCount: Count current file", detail: "Count lines for current file.",  description: null, command: countCurrentFile });
+        items.push({label: "LineCount: Count Workspace files", detail: "Count lines for Workspace files.",  description: null, command: countWorkspace });
+        vscode.window.showQuickPick(items, { matchOnDetail: true, matchOnDescription: true }).then(selectedItem => {
+            if (selectedItem && typeof selectedItem.command === "function") {
+                selectedItem.command();
+            }
+        });               
+    });
+    context.subscriptions.push(disposable3);    
+
+    function configChanged(){
+        counter.getConfig();
+        if(counter.showStatusBarItem)
+            status.show();
+        else
+            status.hide();
+    }    
+    function countCurrentFile(){
+        counter.countCurrentFile();
+    }
+    function countWorkspace(){
+        counter.countWorkspace();
+    }
+
+    configChanged();
 }
 
 // this method is called when your extension is deactivated