+++ CHANGELOG.md
@@ -25,3 +25,11 @@ All notable changes to the "linecount" extension will be documented in this file
 
 
 
+## Version 0.1.4 (2017-02-08)
+
+1. delete continuationmark;
+
+2. add string quotes config.
+
+
+